# backend/spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentAPI Proxy",
    "description": "API for managing agent sessions, schedules, and settings. AgentAPI Proxy provides a unified interface for running AI agents with Kubernetes-based session management.",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/sessions/{sessionId}": {
      "delete": {
        "summary": "Delete a session or pending session allocation",
        "description": "Terminates and deletes a session, or cancels a pending session allocation. Users can only delete their own sessions and allocations.",
        "operationId": "deleteSession",
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session or pending session allocation deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Session terminated successfully"
                    },
                    "session_id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "example": "terminated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Session ID is required"
          },
          "403": {
            "description": "Forbidden - can only delete own sessions"
          },
          "404": {
            "description": "Session not found"
          },
          "409": {
            "description": "Session allocation is no longer pending"
          },
          "500": {
            "description": "Failed to delete session"
          }
        }
      }
    },
...
    "/external-session-managers": {
      "get": {
        "summary": "List External Session Managers",
        "operationId": "listExternalSessionManagers",
        "tags": [
          "External Session Managers"
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "team"
              ]
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registered managers"
          }
        }
      }
    },
...
    "/external-session-managers/enroll": {
      "post": {
        "summary": "Enroll an External Session Manager with a one-time registration token",
        "operationId": "enrollExternalSessionManager",
        "tags": [
          "External Session Managers"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Manager enrolled; connection_token is returned once"
          },
          "401": {
            "description": "Invalid, expired, or already-used registration token"
          }
        }
      }
    },
    "/external-session-managers/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Get an External Session Manager",
        "operationId": "getExternalSessionManager",
        "tags": [
          "External Session Managers"
        ],
        "responses": {
          "200": {
            "description": "Manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalSessionManager"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update an External Session Manager",
        "operationId": "updateExternalSessionManager",
        "tags": [
          "External Session Managers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalSessionManagerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated manager"
          }
        }
      },
      "delete": {
        "summary": "Delete an External Session Manager",
        "operationId": "deleteExternalSessionManager",
        "tags": [
          "External Session Managers"
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
    "/external-session-managers/{id}/rotate-token": {
      "post": {
        "summary": "Rotate an External Session Manager connection token",
        "operationId": "rotateExternalSessionManagerToken",
        "tags": [
          "External Session Managers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token rotated; connection_token is returned once"
          }
        }
      }
    },
    "/external-session-managers/{id}/heartbeat": {
      "post": {
        "summary": "Report native daemon health",
        "description": "Authenticated with the External Session Manager connection token. The parent verifies that public_url/healthz is reachable before accepting the heartbeat.",
        "operationId": "heartbeatExternalSessionManager",
        "tags": [
          "External Session Managers"
        ],
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalSessionManagerHeartbeat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Heartbeat accepted"
          },
          "401": {
            "description": "Invalid connection token"
          },
          "424": {
            "description": "public_url is unreachable"
          }
        }
      }
    },
...
    "/admin/session-managers/{id}": {
      "patch": {
        "summary": "Enable, disable, or drain a session manager", "operationId": "patchSessionManager", "tags": ["Session Pools"],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true}}}},
        "responses": {"200": {"description": "Updated manager"}}
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExternalSessionManagerUpdateRequest": {
        "type": "object",
        "properties": {
          "instance_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "labels": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "default": {
            "type": "boolean"
          },
          "public_url": {
            "type": "string",
            "format": "uri"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "ExternalSessionManager": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "instance_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "has_connection_token": {
            "type": "boolean"
          },
          "connection_token": {
            "type": "string",
            "description": "Returned only on creation or rotation"
          },
          "labels": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "default": {
            "type": "boolean"
          },
          "public_url": {
            "type": "string",
            "format": "uri"
          },
          "version": {
            "type": "string"
          },
          "active_sessions": {
            "type": "integer"
          },
          "last_heartbeat_at": {
            "type": "string",
            "format": "date-time"
          },
          "created": {
            "type": "boolean"
          }
        }
      },
      "ExternalSessionManagerHeartbeat": {
        "type": "object",
        "properties": {
          "public_url": {
            "type": "string",
            "format": "uri"
          },
          "version": {
            "type": "string"
          },
          "active_sessions": {
            "type": "integer"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "GitHub OAuth token"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
...
}